% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_function_helpers.R
\name{get_section}
\alias{get_section}
\title{See a section of a function (for development of patch)}
\usage{
get_section(fbody, tloc)
}
\arguments{
\item{fbody}{Body of a function / (function directly if attached dynamically)}

\item{tloc}{Recursive list path as returned by
\code{\link[=locate_section]{locate_section}} or hand typed}
}
\value{
Language object
}
\description{
See a section of a function (for development of patch)
}
\examples{
# this only happens if user attached optional functions dynamically
if (exists("locate_section") & exists("get_section")) {
  get_section(library, locate_section(library, "logical.return"))
}
}
\seealso{
\code{\link[=locate_section]{locate_section}}
}
\keyword{internal}
