% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_function_helpers.R
\name{locate_section}
\alias{locate_section}
\title{Locate a section of a function}
\usage{
locate_section(fbody, search_str)
}
\arguments{
\item{fbody}{Body of a function / (function directly if attached dynamically)}

\item{search_str}{The string(s) to find in the body}
}
\value{
Recursive list path (if match found) for function body
}
\description{
Locate a section of a function
}
\examples{
# this only happens if user attached optional functions dynamically
if (exists("locate_section") & exists("get_section")) {
  get_section(library, locate_section(library, "logical.return"))
}
}
\seealso{
\code{\link[=get_section]{get_section}}
}
\keyword{internal}
